+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Do not log authorization header in debug traces
+- Fix expand and select of navigations properties
 - Fix error response for errors during media stream read
+- Do not log authorization header in debug traces
+- Decode request url for debug traces
 
 ## Version 1.12.9 - 2024-05-03